requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.5.14
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.104.1
filetype==1.2.0
fonttools==4.55.3
frozenlist==1.5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.63.0rc1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==0.0.11
langchain-text-splitters==0.3.5
langchainplus-sdk==0.0.20
langsmith==0.1.147
marshmallow==3.26.0
matplotlib==3.10.0
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.2
numpy==1.26.4
openai==1.61.0
openapi-schema-pydantic==1.2.4
orjson==3.10.15
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
patsy==1.0.1
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.6
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.0.9
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sse-starlette==1.6.5
starlette==0.27.0
statsmodels==0.14.4
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.24.0
yarl==1.18.3
zstandard==0.23.0
